refactor(transparency): colocate ShapNativeRenderer in shap_visualisers.py
